Separate spec matchers to limit cross-package dependencies
