+++ Specs/customizeJasmine.js
@@ -1,16 +1,27 @@
-import addDefaultMatchers from "./addDefaultMatchers.js";
 import equalsMethodEqualityTester from "./equalsMethodEqualityTester.js";
 
-function customizeJasmine(
-  env,
-  includedCategory,
-  excludedCategory,
-  webglValidation,
-  webglStub,
-  release,
-  debugCanvasWidth,
-  debugCanvasHeight,
-) {
+/**
+ * @param {object} env The jasmine environment.
+ * @param {object} [options] An object with the following properties:
+ * @param {string} [options.includeCategory] Only run specs in this category.
+ * @param {string} [options.excludeCategory] Skip specs in this category.
+ * @param {boolean} [options.webglValidation=false] Enable WebGL validation.
+ * @param {boolean} [options.webglStub=false] Replace WebGL with a stub.
+ * @param {boolean} [options.release=false] True when running against a release build.
+ * @param {number} [options.debugCanvasWidth] The width of the debug canvas.
+ * @param {number} [options.debugCanvasHeight] The height of the debug canvas.
+ */
+function customizeJasmine(env, options = {}) {
+  const {
+    includeCategory,
+    excludeCategory,
+    webglValidation = false,
+    webglStub = false,
+    release = false,
+    debugCanvasWidth,
+    debugCanvasHeight,
+  } = options;
+
   // set this for uniform test resolution across devices
   window.devicePixelRatio = 1;
 
@@ -19,19 +30,17 @@ function customizeJasmine(
   const originalDescribe = window.describe;
 
   window.describe = function (name, suite, category) {
-    // exclude this spec if we're filtering by category and it's not the selected category
-    // otherwise if we have an excluded category, exclude this test if the category of this spec matches
     if (
-      includedCategory &&
-      includedCategory !== "" &&
-      includedCategory !== "none" &&
-      category !== includedCategory
+      includeCategory &&
+      includeCategory !== "" &&
+      includeCategory !== "none" &&
+      category !== includeCategory
     ) {
       window.xdescribe(name, suite);
     } else if (
-      excludedCategory &&
-      excludedCategory !== "" &&
-      category === excludedCategory
+      excludeCategory &&
+      excludeCategory !== "" &&
+      category === excludeCategory
     ) {
       window.xdescribe(name, suite);
     } else {
@@ -51,7 +60,6 @@ function customizeJasmine(
   window.debugCanvasHeight = debugCanvasHeight;
 
   env.beforeEach(function () {
-    addDefaultMatchers(!release).call(env);
     env.addCustomEqualityTester(equalsMethodEqualityTester);
   });
 }
